pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "koda"
version = "0.3.0"
description = "K.O.D.A. — open-source security specialist agent"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Vektra Industries" }]
keywords = ["ai", "agent", "security", "llm", "anthropic", "ollama", "pentest", "soc"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
]
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]
mcp = [
    "mcp>=1.0",
]

[project.scripts]
koda = "koda.cli:main"

[tool.setuptools.packages.find]
include = ["koda*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
# E501: line length handled by formatter, not linter
# E702: compound statements intentional in argv/shift-idiom code
# UP042: StrEnum migration deferred — semantics differ subtly
# B905: zip() strict= is a style lift, not a correctness issue here
ignore = ["E501", "E702", "UP042", "B905"]
